Format

+++ test/adapters/unit/AxelarAdapter.t.sol
@@ -113,7 +113,7 @@ contract AxelarAdapterTest is AxelarAdapterTestBase {
     }
 
     function testEstimate(uint256 gasLimit) public {
-        gasLimit = bound(gasLimit, 1, adapter.RECEIVE_COST() - 1);
+        gasLimit = uint128(bound(gasLimit, 1, adapter.RECEIVE_COST() - 1));
 
         adapter.wire(CENTRIFUGE_CHAIN_ID, abi.encode(AXELAR_CHAIN_ID, REMOTE_AXELAR_ADDR));
